[CL-1941] Move EditModePreview logic to shared folder
